+++ bot/models.py
@@ -7,8 +7,8 @@ from sqlalchemy import create_engine, MetaData, Table, Column, Integer, Unicode
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import Session
 
-from logger import log_print
-from tokens.tokens import DATABASE_HOST
+from bot.logger import log_print
+from bot.tokens.tokens import DATABASE_HOST
 
 
 @contextmanager
