+++ bot/weather.py
@@ -1,9 +1,9 @@
 import pyowm
 from sqlalchemy.orm.exc import NoResultFound
 
-from models import connector, Locations, ENGINE
-from logger import log_print
-from tokens.tokens import ADMINS, WEATHER_TOKEN
+from bot.models import connector, Locations, ENGINE
+from bot.logger import log_print
+from bot.tokens.tokens import ADMINS, WEATHER_TOKEN
 
 
 def weather(bot, update, args):
